fix issue #16
previously, `bool.should` entries were concatenated in `__and__` calls.
Now query A and B are wrapped in a container `bool.must=[A, B]` if
a `should` entry exists in A or B.
If not, the `must`, `must_not` and `filter` entries are still concatenated,
which, i think, is perfectly valid because they all represent AND combinations.